+++ saleor/kitchentransfer/api/views.py
@@ -18,7 +18,7 @@ from .serializers import (
     ItemsSerializer,
     ItemsStockSerializer
      )
-
+from ...decorators import user_trail
 User = get_user_model()
 
 
@@ -26,8 +26,10 @@ class CreateAPIView(generics.CreateAPIView):
     queryset = Table.objects.all()
     serializer_class = CreateListSerializer
 
-    def perform_update(self, serializer):
+    def perform_create(self, serializer):
         serializer.save(user=self.request.user)
+        user_trail(self.request.user.name,
+                   'made a kitchen transfer:#' + str(serializer.data['date']), 'add')
 
 
 class DestroyView(generics.DestroyAPIView):
@@ -37,8 +39,13 @@ class DestroyView(generics.DestroyAPIView):
         items = instance.kitchen_transfer_items.all()
         for item in items:
             Stock.objects.increase_stock(item.stock, item.qty)
-        # raise serializers.ValidationError('You cannot delete ')
-        instance.delete()
+        if instance.any_closed():
+            instance.trashed = True
+            instance.save()
+        else:
+            instance.delete()
+        user_trail(self.request.user.name,
+                   'deleted a kitchen transfer:#', 'add')
 
 
 class DestroyItemView(generics.DestroyAPIView):
@@ -47,7 +54,13 @@ class DestroyItemView(generics.DestroyAPIView):
     def perform_destroy(self, instance):
         Stock.objects.increase_stock(instance.stock, instance.qty)
         # raise serializers.ValidationError('You cannot delete ')
-        instance.delete()
+        if instance.closed:
+            instance.trashed = True
+            instance.save()
+        else:
+            instance.delete()
+        user_trail(self.request.user.name,
+                   'deleted a kitchen transfer item:#', 'add')
 
 
 class ListAPIView(generics.ListAPIView):
@@ -65,13 +78,12 @@ class ListAPIView(generics.ListAPIView):
         return {"date": None, 'request': self.request}
 
     def get_queryset(self, *args, **kwargs):
+        queryset_list = Table.objects.filter(trashed=False)
         try:
             if self.kwargs['pk']:
-                queryset_list = Table.objects.filter(customer__pk=self.kwargs['pk']).order_by('car').distinct('car').select_related()
-            else:
-                queryset_list = Table.objects.all.select_related()
+                queryset_list = queryset_list.filter(customer__pk=self.kwargs['pk']).order_by('car').distinct('car').select_related()
         except Exception as e:
-            queryset_list = Table.objects.all()
+            pass
 
         page_size = 'page_size'
         if self.request.GET.get(page_size):
